requirements.txt
astroid==2.3.2
cachetools==3.1.1
certifi==2019.9.11
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.14.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.4.3
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
Werkzeug==0.16.0
wrapt==1.11.2
